More parser Work, should be close to actaully compiling this damned thing now.

+++ main.cpp
@@ -1,10 +1,9 @@
-#include "TSP.h"
-#include <ostream>
+#include <TSP.hpp>
 
 int main(int argc, char** argv){
 	TSP *TSPGraph = new TSP(argv[0]);
-	for (float* : TSP::list){
-		for (float j : i){
+	for (double* : TSP::list){
+		for (double j : i){
 			std::cout << i << ' ';
 		}
 		std::cout << '/n';
